Moved TTS code using AVFoundation to another file, so tests will compile properly on Linux
